+++ libnodegl/backends/gl/gpu_ctx_gl.h
@@ -77,5 +77,6 @@ struct gpu_ctx_gl {
 int ngli_gpu_ctx_gl_make_current(struct gpu_ctx *s);
 int ngli_gpu_ctx_gl_release_current(struct gpu_ctx *s);
 void ngli_gpu_ctx_gl_reset_state(struct gpu_ctx *s);
+int ngli_gpu_ctx_gl_wrap_framebuffer(struct gpu_ctx *s, GLuint fbo);
 
 #endif