+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnodegl`.
 ### Fixed
 - Various memory leaks
 
+### Added
+- Optional `backend_config` field to the `ngl_config` structure (ABI break)
+- External OpenGL context support
+
 ## [2022.3] [libnodegl 0.3.0] - 2022-04-11
 ### Added
 - `eye`, `center` and `up` Camera parameters can now be node